+++ ICalcify/Explorer.py
@@ -5,9 +5,13 @@ from io import BytesIO
 import json
 from pathlib import Path
 
-from ICalcify.Tree import Tree
+from ICalcify.Tree import read, Tree
 
 m.patch()
+
+def New(fnames,buffer=False):
+    trees = dict([read(f,buffer=buffer,retname=True) for f in fnames])
+    return Explorer(trees)
 
 class Explorer(object):
     def __init__(self, trees):
@@ -18,6 +22,8 @@ class Explorer(object):
 
     def __getitem__(self, key):
         if key in self.trees:
+            if callable(self.trees[key]):
+                self.trees[key] = (self.trees[key])()
             return self.trees[key]
         else:
             raise KeyError("\'{}\' not valid Tree".format(key))
@@ -30,10 +36,24 @@ class Explorer(object):
 
     def __repr__(self):
         out = ""
-        for tree in self.trees:
-            out += "\t{}: {} branches\n".format(tree,len(self.trees[tree]))
+        for key in self.trees:
+            if callable(self.trees[key]):
+                out += "\t{}: *buffered\n".format(key)
+            else:
+                out += "\t{}: {} branches\n".format(key,len(self.trees[key]))
         if len(self) == 0:
             out += '\tEMPTY'
+        return out
+
+    def _repr_html_(self):
+        out = ""
+        for key in self.trees:
+            if callable(self.trees[key]):
+                out += "<h2>{}: &#8250;</h2>".format(key)
+            else:
+                out += "<h2>{}: &#8964;</h2> <div style=\"margin: 20px;\">{}</div>".format(key,self.trees[key]._repr_html_())
+        if len(self) == 0:
+            out += '<h1>EMPTY</h1>'
         return out
 
     def __iter__(self):
@@ -41,3 +61,8 @@ class Explorer(object):
 
     def __len__(self):
         return self.trees.__len__()
+
+    def load_all(self):
+        for key in self.trees:
+            if callable(self.trees[key]):
+                self.trees[key] = (self.trees[key])()