+++ ICalcify/__init__.py
@@ -1,5 +1,5 @@
 from ICalcify.Tree import read, read_jsonc, read_msg, Tree
 from ICalcify.branches import Branch
 from ICalcify import base as IO
-from ICalcify.Explorer import Explorer
+from ICalcify.Explorer import New, Explorer
 from ICalcify.command_line import main