+++ RELEASENOTES.md
@@ -1,7 +1,13 @@
 # Release Notes
 
+### 0.0.5 2020-04-26
+
+* Added buffered reading of Trees into the Explorer
+
+* Added better Jupyter support
+
 ### 0.0.4: 2020-04-21
 
 * Added Explorer class and the ability to read more than one file from the command line.
 
-* Added iterability to Branches for element-wise iteration. And to Explorers and Trees for iterating over keys. 
+* Added iterability to Branches for element-wise iteration. And to Explorers and Trees for iterating over keys.